requirements.txt
colorama>=0.4.6
PyGithub>=2.3.0
python-gitlab>=4.4.0
requests>=2.32.0
